+++ src/packages/__VUE/menu/index.taro.vue
@@ -24,12 +24,14 @@
   </view>
 </template>
 <script lang="ts">
-import { reactive, provide, computed, ref } from 'vue';
+import { PropType, computed, ref } from 'vue';
 import { createComponent } from '@/packages/utils/create';
 import Taro, { usePageScroll } from '@tarojs/taro';
 import { useTaroRect } from '@/packages/utils/useTaroRect';
 const { componentName, create } = createComponent('menu');
 import { RectUp, RectDown } from '@nutui/icons-vue-taro';
+import { useChildren } from '@/packages/utils/useRelation';
+import { MENU_KEY, MenuDirection } from './types';
 
 export default create({
   components: {
@@ -43,11 +45,11 @@ export default create({
     },
     overlay: {
       type: Boolean,
-      default: true as const
+      default: true
     },
     lockScroll: {
       type: Boolean,
-      default: true as const
+      default: true
     },
     duration: {
       type: [Number, String],
@@ -58,14 +60,17 @@ export default create({
       default: true
     },
     direction: {
-      type: String,
+      type: String as PropType<MenuDirection>,
       default: 'down'
     },
     scrollFixed: {
       type: [Boolean, String, Number],
       default: false
     },
-    titleClass: [String]
+    titleClass: {
+      type: String,
+      default: ''
+    }
   },
   setup(props) {
     const barRef = ref<HTMLElement>();
@@ -73,53 +78,7 @@ export default create({
     const offset = ref(0);
     const isScrollFixed = ref(false);
 
-    const useChildren = () => {
-      const publicChildren: any[] = reactive([]);
-      const internalChildren: any[] = reactive([]);
-
-      const linkChildren = (value?: any) => {
-        const link = (child: any) => {
-          if (child.proxy) {
-            internalChildren.push(child);
-            publicChildren.push(child.proxy as any);
-          }
-        };
-
-        const removeLink = (child: any) => {
-          if (child.proxy) {
-            let internalIndex = internalChildren.indexOf(child);
-            if (internalIndex > -1) {
-              internalChildren.splice(internalIndex, 1);
-            }
-
-            let publicIndex = publicChildren.indexOf(child.proxy);
-            if (internalIndex > -1) {
-              publicChildren.splice(publicIndex, 1);
-            }
-          }
-        };
-
-        provide(
-          'menuParent',
-          Object.assign(
-            {
-              removeLink,
-              link,
-              children: publicChildren,
-              internalChildren
-            },
-            value
-          )
-        );
-      };
-
-      return {
-        children: publicChildren,
-        linkChildren
-      };
-    };
-
-    const { children, linkChildren } = useChildren();
+    const { children, linkChildren } = useChildren(MENU_KEY);
 
     const opened = computed(() => children.some((item) => item.state.showWrapper));
