+++ src/packages/__VUE/menu/index.vue
@@ -12,8 +12,8 @@
             <view class="nut-menu__title-text">{{ item.renderTitle() }}</view>
             <span class="nut-menu__title-icon">
               <slot name="icon">
-                <ArrowUp2 v-if="direction === 'up'" />
-                <ArrowDown2 v-else />
+                <RectUp v-if="direction === 'up'" />
+                <RectDown v-else />
               </slot>
             </span>
           </view>
@@ -24,15 +24,17 @@
   </view>
 </template>
 <script lang="ts">
-import { reactive, provide, computed, ref, onMounted, onUnmounted } from 'vue';
+import { computed, ref, onMounted, onUnmounted, PropType } from 'vue';
 import { createComponent } from '@/packages/utils/create';
 import { useRect } from '@/packages/utils/useRect';
 const { componentName, create } = createComponent('menu');
-import { ArrowUp2, ArrowDown2 } from '@nutui/icons-vue';
+import { RectUp, RectDown } from '@nutui/icons-vue';
+import { MENU_KEY, MenuDirection } from './types';
+import { useChildren } from '@/packages/utils/useRelation';
 export default create({
   components: {
-    ArrowUp2,
-    ArrowDown2
+    RectUp,
+    RectDown
   },
   props: {
     activeColor: {
@@ -41,11 +43,11 @@ export default create({
     },
     overlay: {
       type: Boolean,
-      default: true as const
+      default: true
     },
     lockScroll: {
       type: Boolean,
-      default: true as const
+      default: true
     },
     duration: {
       type: [Number, String],
@@ -56,67 +58,24 @@ export default create({
       default: true
     },
     direction: {
-      type: String,
+      type: String as PropType<MenuDirection>,
       default: 'down'
     },
     scrollFixed: {
       type: [Boolean, String, Number],
       default: false
     },
-    titleClass: [String]
+    titleClass: {
+      type: String,
+      default: ''
+    }
   },
   setup(props) {
     const barRef = ref<HTMLElement>();
     const offset = ref(0);
     const isScrollFixed = ref(false);
 
-    const useChildren = () => {
-      const publicChildren: any[] = reactive([]);
-      const internalChildren: any[] = reactive([]);
-
-      const linkChildren = (value?: any) => {
-        const link = (child: any) => {
-          if (child.proxy) {
-            internalChildren.push(child);
-            publicChildren.push(child.proxy as any);
-          }
-        };
-
-        const removeLink = (child: any) => {
-          if (child.proxy) {
-            let internalIndex = internalChildren.indexOf(child);
-            if (internalIndex > -1) {
-              internalChildren.splice(internalIndex, 1);
-            }
-
-            let publicIndex = publicChildren.indexOf(child.proxy);
-            if (internalIndex > -1) {
-              publicChildren.splice(publicIndex, 1);
-            }
-          }
-        };
-
-        provide(
-          'menuParent',
-          Object.assign(
-            {
-              removeLink,
-              link,
-              children: publicChildren,
-              internalChildren
-            },
-            value
-          )
-        );
-      };
-
-      return {
-        children: publicChildren,
-        linkChildren
-      };
-    };
-
-    const { children, linkChildren } = useChildren();
+    const { children, linkChildren } = useChildren(MENU_KEY);
 
     const opened = computed(() => children.some((item) => item.state.showWrapper));
