+++ src/packages/__VUE/menu/types.ts
@@ -1,0 +1,6 @@
+export const MENU_KEY = Symbol('nut-menu');
+export type MenuDirection = 'down' | 'up';
+export type MenuItemOption = {
+  text: string;
+  value: number | string;
+};
