--- src/packages/__VUE/menuitem/type.ts
@@ -1,4 +1,0 @@
-export type MenuItemOption = {
-  text: string;
-  value: number | string;
-};
